move theme instances out of react-theme-provider
