--- packages/react-theme-provider/src/themes/FluentTheme.ts
@@ -1,5 +1,0 @@
-import { Theme } from '../types';
-import { createTheme } from '../createTheme';
-
-// TODO: move this to `@fluentui/theme` package.
-export const FluentTheme: Theme = createTheme();
