+++ packages/storybook/src/themes/v8/FluentTheme.ts
@@ -1,0 +1,3 @@
+import { createTheme, Theme } from '@fluentui/react-theme-provider';
+
+export const FluentTheme: Theme = createTheme();
